pyproject.toml
[tool.poetry]
name = "foodgram-project-react"
version = "0.1.0"
description = ""
authors = ["Vasilii Safronov <vks20001@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
Pillow = "^9.0.1"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
django-filter = "^21.1"
django-extra-fields = "^3.0.2"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
